# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "description": "Product API with Auth + JWT",
        "contact": {
            "email": "hadiatab16@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Demo API Product"
        }
    ],
    "paths": {
...
        "/category-products": {
            "get": {
                "tags": [
                    "Category Products"
                ],
                "summary": "List all category products",
                "operationId": "allCategoryProduct",
                "responses": {
                    "200": {
                        "description": "Category products found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Category Product found",
                                    "data": "Array of category product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category Products"
                ],
                "summary": "Create a new category product",
                "operationId": "storeCategoryProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Category"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Category Product created successfully",
                                    "data": {
                                        "id": 1,
                                        "name": "New Category"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "failed",
                                    "message": "create Category Product failed!",
                                    "errors": {
                                        "name": "array of error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/category-products/{id}": {
            "get": {
                "tags": [
                    "Category Products"
                ],
                "summary": "Get a category product by ID",
                "operationId": "showCategoryProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category Product found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Category Product found",
                                    "data": {
                                        "id": 1,
                                        "name": "HP"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category Product not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "failed",
                                    "message": "Category Product not found"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Category Products"
                ],
                "summary": "Update an existing category product",
                "description": "Update the specified resource in storage.",
                "operationId": "updateCategoryProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Category"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Category Product updated successfully",
                                    "data": {
                                        "id": 1,
                                        "name": "Updated Category"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category Product not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "failed",
                                    "message": "Category Product not found"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "failed",
                                    "message": "update Category Product failed!",
                                    "errors": {
                                        "name": "array of error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category Products"
                ],
                "summary": "Delete a category product",
                "description": "Remove the specified resource from storage.",
                "operationId": "deleteCategoryProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Category Product deleted successfully!"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category Product not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "failed",
                                    "message": "Category Product not found"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token to access these endpoints",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}